test_pth

+++ src/motor.cc
@@ -42,13 +42,13 @@ int move_servo(pantilthatns::pantilthat *pth, size_t servo_id, float angle)
 
 struct pos apply_angle(pantilthatns::pantilthat *pth, struct pos relative_pos)
 {
-    float angle_vertical = pth->get_servo(1);
-    float angle_horizontal = pth->get_servo(2);
+    int angle_vertical = pth->get_servo(1);
+    int angle_horizontal = pth->get_servo(2);
 
-    float new_angle_vertical =
-        ((angle_vertical + (relative_pos.y * 3) + 90) % 180f) - 90;
-    float new_angle_horizontal =
-        ((angle_horizontal + (relative_pos.x * 3) + 90) % 180f) - 90;
+    int new_angle_vertical =
+        ((angle_vertical + (relative_pos.y * 3) + 90) % 180) - 90;
+    int new_angle_horizontal =
+        ((angle_horizontal + (relative_pos.x * 3) + 90) % 180) - 90;
 
     printf("New angle vertical:%f\n", new_angle_vertical);
     printf("New angle horizontal:%f\n", new_angle_horizontal);
